2021년 12월 21일
* 보드FW가 없으므로 python에서 웬만하면 수정하는걸로 진행
* 21, 25버전 비교결과 동일하여 계속 수정해오던 25버전 코드로 계속 수정하겠으

[python]
- 보드 up/down에서 z위치 - totalthickness로 수정한 것 다시 원상복구함 (원래 그렇게 되어있었고, 보드fw에 따라 저렇게 써야할 수도 있으니 일단 장비 테스트 우선 해보고 이상하면 - totalthickness할 것 : 문제 없이 잘 쓰고있다고 하니까...)
- 초기에 opendoor랑 boardhome 찾는다고 while문 돌리는거 그냥 뺌. (굳이 돌릴필요잇나 0점으로 잘 가는거같은데?) 그냥 while 없이 boardhome 으로 가는 코드만 추가 => 안됨, hoardhome으로 가지도 않고 첫 몇층은 올라가기만하고 그럼, 다시 while문 추가함 (boardhome 리턴값 비교하는거 추가)

+++ printer/cls_ML192v22.py
@@ -32,7 +32,7 @@ class cls_ML192v22(object):
         self.LightTime_Beginning = float(config.get('PARAMETER', 'LightTime_Beginning'))
         self.DivisionTime = float(config.get('PARAMETER', 'DivisionTime'))
         self.ZMotorEndUpDistance = int(config.get('MOTOR_SETUP', 'ZMotorEndUpDistance'))
-        self.ZMotorAmount = float(config.get('MOTOR_SETUP', 'ZMotorAmount'))
+        self.ZMotorAmount = int(config.get('MOTOR_SETUP', 'ZMotorAmount'))
         self.ZeroPoint = float(config.get('MOTOR_SETUP', 'ZeroPoint'))
         # self.WaitTime_Basic = float(config.get('PARAMETER', 'BasicWaitTime'))
         # self.WaitTime_Beginning = float(config.get('PARAMETER', 'InitWaitTime'))
